requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
SQLAlchemy==2.0.46
Werkzeug==3.1.6
python-dotenv==1.0.0
